requirements.txt
Flask
pandas
scikit-learn
nltk
pyspellchecker
googletrans==4.0.0-rc1
langdetect